combine estimation tool page queries to match convention
